refactor: common/store/service 패키지구조 변경 및 service interface 생성
