+++ common/src/main/java/com/food/common/store/business/StoreOwnerDomainService.java
@@ -1,0 +1,10 @@
+package com.food.common.store.business;
+
+import com.food.common.store.domain.StoreOwner;
+import com.food.common.user.domain.User;
+
+import java.util.Optional;
+
+public interface StoreOwnerDomainService {
+    Optional<StoreOwner> findByUser(User user);
+}
