+++ common/src/main/java/com/food/common/store/business/impl/DefaultStoreOwnerDomainService.java
@@ -1,5 +1,6 @@
-package com.food.common.store.service.storeOwner;
+package com.food.common.store.business.impl;
 
+import com.food.common.store.business.StoreOwnerDomainService;
 import com.food.common.store.domain.StoreOwner;
 import com.food.common.user.domain.User;
 import com.food.common.user.repository.StoreOwnerRepository;
@@ -10,7 +11,7 @@ import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
-public class StoreOwnerDomainService {
+public class DefaultStoreOwnerDomainService implements StoreOwnerDomainService {
     private final StoreOwnerRepository storeOwnerRepository;
 
     public Optional<StoreOwner> findByUser(User user) {
